Update kalkuliatorius.py
Pakeiciau int i float, kad jei bus naudojami skaiciai su kableliu.

+++ kalkuliatorius.py
@@ -30,8 +30,8 @@ while True:
     # Patikrinimas, ar tai yra vienas is pasirinkimu
     if pasirinkimas in ('1', '2', '3', '4'):
         try:
-            pirmas = int(intput("Pirmas numeris:"))
-            antras = int(intput("Antras numeris:"))
+            pirmas = float(intput("Pirmas numeris:"))
+            antras = float(intput("Antras numeris:"))
         except ValueError:
             print("Netinkamas. Prasom irasyti numeri")
             continue
